+++ src/components/encryptor.js
@@ -51,6 +51,20 @@ export const Encryptor = ({ library, context, publicKey }) => {
     },
 
     /**
+     * Delete the underlying WASM instance
+     *
+     * Should be called before dereferencing this object
+     * @function
+     * @name IEncryptor#delete
+     */
+    delete() {
+      if (_instance) {
+        _instance.delete()
+        _instance = null
+      }
+    },
+
+    /**
      * Encrypts a PlainText and stores the result in the destination parameter.
      * Dynamic memory allocations in the process are allocated from the memory
      * pool pointed to by the given MemoryPoolHandle.